requirements.txt
alabaster==0.7.12
asgiref==3.4.1
autopep8==1.5.6
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bootstrap-py==1.0.2
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-froala-editor==3.2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-quill-editor==0.1.22
django-summernote==0.8.11.6
djangorestframework==3.12.4
docutils==0.17.1
et-xmlfile==1.0.1
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
humanize==3.12.0
idna==2.10
imagesize==1.2.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
openpyxl==3.0.6
packaging==21.0
parso==0.8.2
pguard==0.4.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2==2.9.1
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.5
python-dotenv==0.15.0
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2
Sphinx==4.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==2.0.1
django-tinymce==3.3.0
